Move sequelize connection to index.js
