+++ collab-service/index.js
@@ -10,6 +10,40 @@ import {
   switchRoles,
   setQuestion,
 } from "./controller/room-controller.js";
+import RoomModelSchema from "./model/room-model.js";
+import "dotenv/config";
+import { Sequelize, DataTypes } from "sequelize";
+
+// Set up sequelize connection
+async function initSequelize() {
+  const sequelizeHost = process.env.NODE_ENV == "production" ? process.env.DB_CLOUD_HOST : "localhost";
+  const sequelizeStorage = process.env.NODE_ENV == "production"
+                            ? process.env.DB_CLOUD_STORAGE
+                            : process.env.NODE_ENV == "development"
+                              ? "./db/roomDB.sqlite"
+                              : "./db/roomDB-test.sqlite";
+
+  let sequelize = new Sequelize("database", "username", "password", {
+    host: sequelizeHost,
+    dialect: "sqlite",
+    storage: sequelizeStorage,
+  });
+
+  let Room;
+
+  try {
+    await sequelize.authenticate();
+    console.log("Connection with SQLite has been established!");
+    Room = RoomModelSchema(sequelize, DataTypes);
+    Room.sync({ force: true });
+  } catch (err) {
+    console.error("Unable to connect to room DB :(");
+  }
+
+  return { sequelize: sequelize, Room: Room };
+}
+
+export const { sequelize, Room } = await initSequelize();
 
 const app = express();
 app.use(express.urlencoded({ extended: true }));
